+++ src/main/java/woowacourse/shoppingcart/application/AuthService.java
@@ -3,11 +3,11 @@ package woowacourse.shoppingcart.application;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import woowacourse.shoppingcart.dto.LoginRequest;
-import woowacourse.shoppingcart.support.JwtTokenProvider;
 import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.dto.LoginRequest;
 import woowacourse.shoppingcart.exception.badrequest.InvalidLoginException;
 import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
+import woowacourse.shoppingcart.support.JwtTokenProvider;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
