+++ src/test/java/woowacourse/shoppingcart/application/unit/AuthServiceTest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.application;
+package woowacourse.shoppingcart.application.unit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.springframework.security.crypto.bcrypt.BCrypt;
+import woowacourse.shoppingcart.application.AuthService;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.LoginRequest;
 import woowacourse.shoppingcart.exception.badrequest.InvalidLoginException;
