+++ src/test/java/woowacourse/shoppingcart/application/unit/ServiceMockTest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.application;
+package woowacourse.shoppingcart.application.unit;
 
 import static org.mockito.Mockito.mockStatic;
 
@@ -9,8 +9,9 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.bcrypt.BCrypt;
-import woowacourse.shoppingcart.support.JwtTokenProvider;
+import woowacourse.shoppingcart.application.CustomerService;
 import woowacourse.shoppingcart.dao.CustomerDao;
+import woowacourse.shoppingcart.support.JwtTokenProvider;
 
 @ExtendWith(MockitoExtension.class)
 abstract class ServiceMockTest {
