+++ src/test/java/woowacourse/shoppingcart/ui/ControllerTest.java
@@ -8,9 +8,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 import woowacourse.shoppingcart.application.AuthService;
-import woowacourse.shoppingcart.support.JwtTokenProvider;
 import woowacourse.shoppingcart.application.CustomerService;
 import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.support.JwtTokenProvider;
 
 @WebMvcTest({
         AuthController.class,
